Create KMS keys if needed
Rather than rely on aws/rds, which isn't guaranteed to exist (it's
created in each account/region when you create an RDS resource with
encryption and don't specify an explicit KMS key).

Create separate keys for storage and performance insights, adding a
hardcoded policy to both (copied from aws/rds). Output the key ARNs as
well.

Finally, split KMS logic out to a separate file, because the policy is
so large it makes main.tf hard to read otherwise.