feat!: allow null values for backup_window and retention_period
This sets the default for backup_window and retention_period to `null`.

When AWS backup is used, the settings either have to match the AWS backup settings exactly or they
have to be unset.

Trying to apply any diverging settings will result in an error:

```
InvalidParameterValue: RDS cluster $name is associated with the following AwsBackupRecoveryPointArn: arn:aws:backup:$region:$account:recovery-point:$mode:cluster-$id. The BackupRetentionPeriod can be blank, or you can use the current value, $current_value. For more details, see the AWS Backup documentation
```

Since matching these settings manually defeats the purpose of having a single source of truth,
this change updates the default values to be `null`.

I would prefer to make this a non-breaking change, but without adding another variable, this seems to not be possible
since a default can't be overridden with `null`.

BREAKING CHANGE: backup_window and retention_period are now `null` by default. If you want to keep the old
default configuration, set `backup_window = "07:00-09:00"` and `retention_period = 5`.